+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 ## Table Of Contents
 
+* [0.5.0](#0.5.0)
 * [0.1.0](#0.1.0)
 
-## 0.0.1
+## 0.5.0
+
+- Added Success Block.
+- Added Progress Block.
+- Added Failure Block.
+
+## 0.1.0
 
 - Download task that can be paused and resumed.
 - Session can be configured to allow multiple concurrent downloads.